Reorganize project layout into modules/, config/, launchers/
Move non-entry-point Python files into modules/, data/config files into
config/, and generated launcher scripts into launchers/. The main entry
point llamacpp-server-launcher.py stays at repo root.

- modules/: about_tab, config, env_vars_module, ik_llama, launch, system
- config/: chat_templates.json, llama_cpp_launcher_configs.json, version
- launchers/: destination for generated .sh / .ps1 scripts
- .gitignore updated for new paths; launchers/.gitkeep keeps folder tracked
- One-time migration: a populated llama_cpp_launcher_configs.json at the
  old repo root is moved into config/ on first run so existing users do
  not lose their saved configurations
